move dsl package
