+++ src/main/groovy/org/boozallen/plugins/jte/init/PipelineDecorator.groovy
@@ -17,8 +17,8 @@ package org.boozallen.plugins.jte.init
 
 import hudson.ExtensionList
 import hudson.model.InvisibleAction
-import org.boozallen.plugins.jte.init.dsl.PipelineConfigurationDsl
-import org.boozallen.plugins.jte.init.dsl.PipelineConfigurationObject
+import org.boozallen.plugins.jte.init.governance.config.dsl.PipelineConfigurationDsl
+import org.boozallen.plugins.jte.init.governance.config.dsl.PipelineConfigurationObject
 import org.boozallen.plugins.jte.init.governance.GovernanceTier
 import org.boozallen.plugins.jte.init.governance.config.ScmPipelineConfigurationProvider
 import org.boozallen.plugins.jte.init.primitives.TemplateBinding
