+++ src/main/groovy/org/boozallen/plugins/jte/init/governance/config/ConsolePipelineConfigurationProvider.groovy
@@ -17,8 +17,8 @@ package org.boozallen.plugins.jte.init.governance.config
 
 import hudson.Extension
 import hudson.Util
-import org.boozallen.plugins.jte.init.dsl.PipelineConfigurationDsl
-import org.boozallen.plugins.jte.init.dsl.PipelineConfigurationObject
+import org.boozallen.plugins.jte.init.governance.config.dsl.PipelineConfigurationDsl
+import org.boozallen.plugins.jte.init.governance.config.dsl.PipelineConfigurationObject
 import org.jenkinsci.plugins.workflow.flow.FlowExecutionOwner
 import org.kohsuke.stapler.DataBoundConstructor
