+++ src/main/groovy/org/boozallen/plugins/jte/init/governance/config/NullPipelineConfigurationProvider.groovy
@@ -16,7 +16,7 @@
 package org.boozallen.plugins.jte.init.governance.config
 
 import hudson.Extension
-import org.boozallen.plugins.jte.init.dsl.PipelineConfigurationObject
+import org.boozallen.plugins.jte.init.governance.config.dsl.PipelineConfigurationObject
 import org.jenkinsci.plugins.workflow.flow.FlowExecutionOwner
 import org.kohsuke.stapler.DataBoundConstructor
