+++ src/main/groovy/org/boozallen/plugins/jte/init/governance/config/ScmPipelineConfigurationProvider.groovy
@@ -19,8 +19,8 @@ import hudson.Extension
 import hudson.Util
 import hudson.scm.NullSCM
 import hudson.scm.SCM
-import org.boozallen.plugins.jte.init.dsl.PipelineConfigurationDsl
-import org.boozallen.plugins.jte.init.dsl.PipelineConfigurationObject
+import org.boozallen.plugins.jte.init.governance.config.dsl.PipelineConfigurationDsl
+import org.boozallen.plugins.jte.init.governance.config.dsl.PipelineConfigurationObject
 import org.boozallen.plugins.jte.util.FileSystemWrapper
 import org.boozallen.plugins.jte.util.TemplateLogger
 import org.jenkinsci.plugins.workflow.flow.FlowExecutionOwner
