+++ src/main/groovy/org/boozallen/plugins/jte/init/governance/config/dsl/TemplateConfigException.groovy
@@ -13,7 +13,7 @@
     See the License for the specific language governing permissions and
     limitations under the License.
 */
-package org.boozallen.plugins.jte.init.dsl
+package org.boozallen.plugins.jte.init.governance.config.dsl
 
 import groovy.transform.InheritConstructors
