+++ src/main/groovy/org/boozallen/plugins/jte/init/primitives/TemplatePrimitiveInjector.groovy
@@ -18,7 +18,7 @@ package org.boozallen.plugins.jte.init.primitives
 import hudson.ExtensionList
 import hudson.ExtensionPoint
 import jenkins.model.Jenkins
-import org.boozallen.plugins.jte.init.dsl.PipelineConfigurationObject
+import org.boozallen.plugins.jte.init.governance.config.dsl.PipelineConfigurationObject
 import org.jenkinsci.plugins.workflow.flow.FlowExecutionOwner
 
 /**
