+++ src/main/groovy/org/boozallen/plugins/jte/init/primitives/hooks/HookInjector.groovy
@@ -17,7 +17,7 @@ package org.boozallen.plugins.jte.init.primitives.hooks
 
 import hudson.Extension
 import jenkins.model.Jenkins
-import org.boozallen.plugins.jte.init.dsl.PipelineConfigurationObject
+import org.boozallen.plugins.jte.init.governance.config.dsl.PipelineConfigurationObject
 import org.boozallen.plugins.jte.init.primitives.TemplatePrimitiveInjector
 import org.boozallen.plugins.jte.util.TemplateScriptEngine
 import org.jenkinsci.plugins.workflow.flow.FlowExecutionOwner
