+++ src/main/resources/org/boozallen/plugins/jte/init/primitives/injectors/ApplicationEnvironment.groovy
@@ -16,7 +16,7 @@
 
 package org.boozallen.plugins.jte.init.primitives.injectors
 
-import org.boozallen.plugins.jte.init.dsl.TemplateConfigException
+import org.boozallen.plugins.jte.init.governance.config.dsl.TemplateConfigException
 import org.boozallen.plugins.jte.init.primitives.TemplateException
 import org.boozallen.plugins.jte.init.primitives.TemplatePrimitive
