+++ src/test/groovy/org/boozallen/plugins/jte/init/governance/config/dsl/PipelineConfigurationObjectSpec.groovy
@@ -13,7 +13,7 @@
     See the License for the specific language governing permissions and
     limitations under the License.
 */
-package org.boozallen.plugins.jte.init.dsl
+package org.boozallen.plugins.jte.init.governance.config.dsl
 
 import hudson.model.TaskListener
 import org.boozallen.plugins.jte.util.TestUtil
